Update Readme and Commands
