+++ ricaun.Revit.UI.StatusBar/Revit/Commands/CommandRevit.cs
@@ -1,6 +1,7 @@
 ﻿using Autodesk.Revit.Attributes;
 using Autodesk.Revit.DB;
 using Autodesk.Revit.UI;
+using ricaun.Revit.UI.StatusBar.Utils;
 
 namespace ricaun.Revit.UI.StatusBar.Revit.Commands
 {
@@ -23,4 +24,5 @@ namespace ricaun.Revit.UI.StatusBar.Revit.Commands
             return Result.Succeeded;
         }
     }
+
 }
