+++ ricaun.Revit.UI.StatusBar/ricaun.Revit.UI.StatusBar.csproj
@@ -16,7 +16,7 @@
 
   <!-- RevitVersion -->
   <PropertyGroup>
-    <TargetFrameworks>net47;net7.0-windows;net8.0-windows</TargetFrameworks>
+    <TargetFrameworks>net47;net8.0-windows</TargetFrameworks>
   </PropertyGroup>
   <Choose>
     <When Condition="$(TargetFramework.StartsWith('net4'))">
